更新tools编译选项 URL: https://qingluan.hozonauto.com/geelib/repoMergeDetail?repoId=algorithm/mapping&mergeId=694

+++ tools/compile.py
@@ -192,8 +192,10 @@ def x86_build(workspace, platform, build_directory, release_directory, **kwargs)
         args[pkg_cmake_enable] = 'ON' if kwargs[pkg] else "OFF"
     # args['-DENABLE_COMPILE_BASE'] = 'ON' if kwargs['base'] else "OFF"
     if kwargs['proto']:
-        execute_shell("rm depend/third_party/cmake/FindProtobuf.cmake")
-        execute_shell("mv depend/third_party/cmake/FindProtobuf.cmakebk depend/third_party/cmake/FindProtobuf.cmake")
+        backup_file = "depend/third_party/cmake/FindProtobuf.cmakebk"
+        if os.path.exists(backup_file):
+            execute_shell("rm depend/third_party/cmake/FindProtobuf.cmake")
+            execute_shell("mv depend/third_party/cmake/FindProtobuf.cmakebk depend/third_party/cmake/FindProtobuf.cmake")
 
     if kwargs['ccache']:
         args['-DCMAKE_C_COMPILER_LAUNCHER'] = 'ccache'
@@ -240,8 +242,10 @@ def orin_build(workspace, platform, build_directory, release_directory, **kwargs
         args[pkg_cmake_enable] = 'ON' if kwargs[pkg] else "OFF"
     # args['-DENABLE_COMPILE_BASE'] = 'ON' if kwargs['base'] else "OFF"
     if kwargs['proto']:
-        execute_shell("rm depend/third_party/cmake/FindProtobuf.cmake")
-        execute_shell("mv depend/third_party/cmake/FindProtobuf.cmakebk depend/third_party/cmake/FindProtobuf.cmake")
+        backup_file = "depend/third_party/cmake/FindProtobuf.cmakebk"
+        if os.path.exists(backup_file):
+            execute_shell("rm depend/third_party/cmake/FindProtobuf.cmake")
+            execute_shell("mv depend/third_party/cmake/FindProtobuf.cmakebk depend/third_party/cmake/FindProtobuf.cmake")
 
     if kwargs['ccache']:
         args['-DCMAKE_C_COMPILER_LAUNCHER'] = 'ccache'
